Issue #618 Rename files
